don't pull in asset.dart file
don't pull in asset.dart file when compiling the dart worker
