+++ bindings/dart/lib/package.dart
@@ -1,7 +1,6 @@
 export 'src/common/data/document.dart' show Document;
 export 'src/common/data/history.dart'
     show UserFeedback, History, Relevance, DayOfWeek, UserAction;
-export 'src/common/reranker/ai.dart' show RerankMode;
 export 'src/common/reranker/ai.dart'
     if (dart.library.io) 'src/mobile/reranker/ai.dart'
     if (dart.library.js) 'src/web/reranker/ai.dart' show XaynAi;
@@ -13,4 +12,5 @@ export 'src/common/reranker/data_provider.dart'
     if (dart.library.js) 'src/web/reranker/data_provider.dart'
     show getAssets, SetupData;
 export 'src/common/reranker/debug.dart' show RerankDebugCallData;
+export 'src/common/reranker/mode.dart' show RerankMode;
 export 'src/common/result/outcomes.dart' show RerankingOutcomes;
