+++ bindings/dart/lib/src/common/reranker/ai.dart
@@ -1,50 +1,14 @@
-import 'dart:math' show max, min;
 import 'dart:typed_data' show Uint8List;
 
-import 'package:json_annotation/json_annotation.dart' show JsonValue;
 import 'package:xayn_ai_ffi_dart/src/common/data/document.dart' show Document;
 import 'package:xayn_ai_ffi_dart/src/common/data/history.dart' show History;
-import 'package:xayn_ai_ffi_dart/src/common/ffi/genesis.dart' as ffi
-    show RerankMode;
 import 'package:xayn_ai_ffi_dart/src/common/reranker/analytics.dart'
     show Analytics;
 import 'package:xayn_ai_ffi_dart/src/common/reranker/data_provider.dart'
     show SetupData;
+import 'package:xayn_ai_ffi_dart/src/common/reranker/mode.dart' show RerankMode;
 import 'package:xayn_ai_ffi_dart/src/common/result/outcomes.dart'
     show RerankingOutcomes;
-
-/// Rerank mode
-enum RerankMode {
-  @JsonValue(ffi.RerankMode.StandardNews)
-  standardNews,
-  @JsonValue(ffi.RerankMode.PersonalizedNews)
-  personalizedNews,
-  @JsonValue(ffi.RerankMode.StandardSearch)
-  standardSearch,
-  @JsonValue(ffi.RerankMode.PersonalizedSearch)
-  personalizedSearch,
-}
-
-extension RerankModeToInt on RerankMode {
-  /// Gets the discriminant.
-  int toInt() {
-    // We can't use `_$RerankModeEnumMap` as it only gets generated for
-    // files which have a `@JsonSerializable` type containing the enum.
-    // You can't make enums `@JsonSerializable`. Given that `RerankMode`
-    // has only few variants and rarely changes we just write this switch
-    // statement by hand.
-    switch (this) {
-      case RerankMode.standardNews:
-        return ffi.RerankMode.StandardNews;
-      case RerankMode.personalizedNews:
-        return ffi.RerankMode.PersonalizedNews;
-      case RerankMode.standardSearch:
-        return ffi.RerankMode.StandardSearch;
-      case RerankMode.personalizedSearch:
-        return ffi.RerankMode.PersonalizedSearch;
-    }
-  }
-}
 
 /// The Xayn AI.
 class XaynAi {
@@ -98,15 +62,3 @@ class XaynAi {
   /// Frees the memory.
   Future<void> free() async => throw UnsupportedError('Unsupported platform.');
 }
-
-/// Maximum number of threads to be used for multithreaded features.
-const int maxNumberOfThreads = 16;
-
-/// Selects the number of threads used by the [`XaynAi`] thread pool.
-///
-/// On a single core system the thread pool consists of only one thread.
-/// On a multicore system the thread pool consists of
-/// (the number of logical cores - 1) threads, but at most [`maxNumberOfThreads`]
-/// threads and at least one thread.
-int selectThreadPoolSize(int numberOfProcessors) =>
-    min(max(numberOfProcessors - 1, 1), maxNumberOfThreads);
