+++ bindings/dart/lib/src/web/ffi/library.dart
@@ -2,9 +2,11 @@
 library library;
 
 import 'dart:html' show WorkerGlobalScope;
+
 import 'package:js/js.dart' show JS;
 import 'package:js/js_util.dart' show promiseToFuture;
-import 'package:xayn_ai_ffi_dart/src/common/reranker/ai.dart'
+
+import 'package:xayn_ai_ffi_dart/src/common/reranker/utils.dart'
     show selectThreadPoolSize;
 
 @JS('Promise')
