+++ bindings/dart/lib/src/web/worker/oneshot.dart
@@ -1,6 +1,7 @@
 import 'dart:html' show MessageChannel, MessageEvent, MessagePort;
 
 import 'package:json_annotation/json_annotation.dart' show JsonSerializable;
+
 import 'package:xayn_ai_ffi_dart/src/common/utils.dart' show ToJson;
 import 'package:xayn_ai_ffi_dart/src/web/worker/message/utils.dart'
     show MessagePortConverter;
